Create README.md
I love learning new technologies and try to be familiar with them in short period of time. Working with teammates is always a privilege  for me. Working as a community helps us achieve dreadful task easily. My thought is that when we grow as a community we will be able to give something big, which is not possible as an individual.